requirements.txt
# Core dependencies - COMPATIBLE VERSIONS FOR PYTHON 3.13
streamlit==1.28.0
pandas==2.0.3  # Downgrade for Python 3.13 compatibility
numpy==1.24.0

# Web Scraping
selenium==4.15.0
undetected-chromedriver==3.5.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.0
requests==2.31.0
aiohttp==3.9.0

# Machine Learning - USE COMPATIBLE VERSIONS
transformers==4.34.0
torch==2.0.1
spacy==3.6.0
textblob==0.17.1
vaderSentiment==3.3.2
scikit-learn==1.3.0

# Visualization
plotly==5.15.0
matplotlib==3.7.0
seaborn==0.12.2
wordcloud==1.9.2

# Database & Utilities
sqlalchemy==2.0.0
python-dotenv==1.0.0
tqdm==4.65.0
pillow==10.0.0

# Development
pytest==7.4.0
black==23.0.0
flake8==6.0.0